requirements.txt
fastapi==0.115.8
Requests==2.32.3
uvicorn==0.34.0
Jinja2==3.1.5
python-dotenv==1.0.0  # 添加 python-dotenv 库
pydantic-settings==2.7.1